+++ back/src/main/java/fr/centralesupelec/thuv/dtos/LogResponseDto.java
@@ -1,0 +1,12 @@
+package fr.centralesupelec.thuv.dtos;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class LogResponseDto {
+    private String name;
+    private String image;
+    private List<TaskLogDto> tasks;
+}